Move some pipeline tests to copy_pipeline
