+++ builder/test/pipeline.cc
@@ -1,8 +1,5 @@
 #include <gtest/gtest.h>
 
-#include <algorithm>
-#include <cassert>
-#include <fstream>
 #include <numeric>
 
 #include "builder/pipeline.h"
@@ -118,19 +115,6 @@ public:
   }
 };
 
-// init_random() for raw_buffer requires allocation be done by caller
-template <typename T>
-void fill_random(raw_buffer& x) {
-  assert(x.base != nullptr);
-  for_each_index(x, [&](auto i) { *reinterpret_cast<T*>(x.address_at(i)) = (rand() % 20) - 10; });
-}
-
-template <typename T, std::size_t N>
-void init_random(buffer<T, N>& x) {
-  x.allocate();
-  fill_random<T>(x);
-}
-
 // Matrix multiplication (not fast!)
 template <typename T>
 index_t matmul(const buffer<const T>& a, const buffer<const T>& b, const buffer<T>& c) {
@@ -151,37 +135,6 @@ index_t matmul(const buffer<const T>& a, const buffer<const T>& b, const buffer<
     }
   }
   return 0;
-}
-
-// A 2D stencil, sums [x + dx0, x + dx1] x [y + dy0, y + dy]
-template <typename T, int dx0, int dy0, int dx1, int dy1>
-index_t sum_stencil(const buffer<const T>& in, const buffer<T>& out) {
-  assert(in.rank == 2);
-  assert(out.rank == 2);
-  for (index_t y = out.dim(1).begin(); y < out.dim(1).end(); ++y) {
-    for (index_t x = out.dim(0).begin(); x < out.dim(0).end(); ++x) {
-      T sum = 0;
-      for (index_t dy = dy0; dy <= dy1; ++dy) {
-        for (index_t dx = dx0; dx <= dx1; ++dx) {
-          sum += in(x + dx, y + dy);
-        }
-      }
-      out(x, y) = sum;
-    }
-  }
-  return 0;
-}
-
-// A centered 2D 3x3 stencil operation.
-template <typename T>
-index_t sum3x3(const buffer<const T>& in, const buffer<T>& out) {
-  return sum_stencil<T, -1, -1, 1, 1>(in, out);
-}
-
-// A centered 2D 5x5 stencil operation.
-template <typename T>
-index_t sum5x5(const buffer<const T>& in, const buffer<T>& out) {
-  return sum_stencil<T, -2, -2, 2, 2>(in, out);
 }
 
 const auto loop_modes = testing::Values(loop::serial, loop::parallel);
@@ -859,102 +812,6 @@ TEST_P(stencil_chain, pipeline) {
   }
 }
 
-TEST(flip_y, pipeline) {
-  // Make the pipeline
-  node_context ctx;
-
-  auto in = buffer_expr::make(ctx, "in", 2, sizeof(char));
-  auto out = buffer_expr::make(ctx, "out", 2, sizeof(char));
-  auto intm = buffer_expr::make(ctx, "intm", 2, sizeof(char));
-
-  var x(ctx, "x");
-  var y(ctx, "y");
-
-  func copy = func::make(copy_2d<char>, {{in, {point(x), point(y)}}}, {{intm, {x, y}}});
-  func flip = func::make(flip_y<char>, {{intm, {point(x), point(-y)}}}, {{out, {x, y}}});
-
-  pipeline p = build_pipeline(ctx, {in}, {out});
-
-  // Run the pipeline.
-  const int H = 20;
-  const int W = 10;
-  buffer<char, 2> in_buf({W, H});
-  init_random(in_buf);
-
-  buffer<char, 2> out_buf({W, H});
-  out_buf.dim(1).translate(-H + 1);
-  out_buf.allocate();
-  const raw_buffer* inputs[] = {&in_buf};
-  const raw_buffer* outputs[] = {&out_buf};
-  test_context eval_ctx;
-  p.evaluate(inputs, outputs, eval_ctx);
-  ASSERT_EQ(eval_ctx.heap.total_size, W * H * sizeof(char));
-  ASSERT_EQ(eval_ctx.heap.total_count, 1);
-
-  for (int y = 0; y < H; ++y) {
-    for (int x = 0; x < W; ++x) {
-      ASSERT_EQ(out_buf(x, -y), in_buf(x, y));
-    }
-  }
-}
-
-TEST(padded_copy, pipeline) {
-  // Make the pipeline
-  node_context ctx;
-
-  auto in = buffer_expr::make(ctx, "in", 2, sizeof(char));
-  auto out = buffer_expr::make(ctx, "out", 2, sizeof(char));
-  auto intm = buffer_expr::make(ctx, "intm", 2, sizeof(char));
-
-  var x(ctx, "x");
-  var y(ctx, "y");
-
-  // We could just clamp using the bounds directly below, but that would hardcode the bounds we clamp
-  // in the pipeline. This way, the bounds can vary at eval-time.
-  var w(ctx, "w");
-  var h(ctx, "h");
-
-  // Copy the input so we can measure the size of the buffer we think we need internally.
-  func copy = func::make(copy_2d<char>, {{in, {point(x), point(y)}}}, {{intm, {x, y}}});
-  // This is elementwise, but with a clamp to limit the bounds required of the input.
-  func crop = func::make(
-      zero_padded_copy<char>, {{intm, {point(clamp(x, 0, w - 1)), point(clamp(y, 0, h - 1))}}}, {{out, {x, y}}});
-
-  crop.loops({y});
-
-  pipeline p = build_pipeline(ctx, {w, h}, {in}, {out});
-
-  const int W = 8;
-  const int H = 5;
-
-  // Run the pipeline.
-  buffer<char, 2> in_buf({W, H});
-  init_random(in_buf);
-
-  // Ask for an output padded in every direction.
-  buffer<char, 2> out_buf({W * 3, H * 3});
-  out_buf.translate(-W, -H);
-  out_buf.allocate();
-
-  index_t args[] = {W, H};
-  const raw_buffer* inputs[] = {&in_buf};
-  const raw_buffer* outputs[] = {&out_buf};
-  test_context eval_ctx;
-  p.evaluate(args, inputs, outputs, eval_ctx);
-  ASSERT_EQ(eval_ctx.heap.total_size, W * H * sizeof(char));
-  ASSERT_EQ(eval_ctx.heap.total_count, 1);
-
-  for (int y = -H; y < 2 * H; ++y) {
-    for (int x = -W; x < 2 * W; ++x) {
-      if (0 <= x && x < W && 0 <= y && y < H) {
-        ASSERT_EQ(out_buf(x, y), in_buf(x, y));
-      } else {
-        ASSERT_EQ(out_buf(x, y), 0);
-      }
-    }
-  }
-}
-
 class multiple_outputs : public testing::TestWithParam<std::tuple<int, int>> {};
 
 INSTANTIATE_TEST_SUITE_P(split_mode, multiple_outputs, testing::Combine(loop_modes, testing::Range(0, 4)),
@@ -1176,256 +1033,6 @@ TEST(unrelated, pipeline) {
   for (int i = 0; i < N2; ++i) {
     ASSERT_EQ(out2_buf(i), 2 * i + 1);
   }
-}
-
-class copied_result : public testing::TestWithParam<int> {};
-
-INSTANTIATE_TEST_SUITE_P(schedule, copied_result, testing::Range(0, 3));
-
-TEST_P(copied_result, pipeline) {
-  int schedule = GetParam();
-
-  // Make the pipeline
-  node_context ctx;
-
-  auto in = buffer_expr::make(ctx, "in", 2, sizeof(short));
-  auto out = buffer_expr::make(ctx, "out", 2, sizeof(short));
-
-  auto intm = buffer_expr::make(ctx, "intm", 2, sizeof(short));
-
-  var x(ctx, "x");
-  var y(ctx, "y");
-
-  // In this pipeline, the result is copied to the output. We should just compute the result directly in the output.
-  func stencil = func::make(sum3x3<short>, {{in, {bounds(-1, 1) + x, bounds(-1, 1) + y}}}, {{intm, {x, y}}});
-  func padded = func::make_copy({intm, {point(x), point(y)}}, {out, {x, y}});
-
-  switch (schedule) {
-  case 0: break;
-  case 1:
-    padded.loops({y});
-    stencil.compute_root();
-    break;
-  case 2: padded.loops({y}); break;
-  }
-
-  pipeline p = build_pipeline(ctx, {in}, {out});
-
-  // Run the pipeline.
-  const int W = 20;
-  const int H = 10;
-  buffer<short, 2> in_buf({W + 2, H + 2});
-  in_buf.translate(-1, -1);
-  buffer<short, 2> out_buf({W, H});
-
-  init_random(in_buf);
-  out_buf.allocate();
-
-  // Not having span(std::initializer_list<T>) is unfortunate.
-  const raw_buffer* inputs[] = {&in_buf};
-  const raw_buffer* outputs[] = {&out_buf};
-  test_context eval_ctx;
-  p.evaluate(inputs, outputs, eval_ctx);
-  ASSERT_EQ(eval_ctx.heap.total_count, 0);
-
-  for (int y = 0; y < H; ++y) {
-    for (int x = 0; x < W; ++x) {
-      int correct = 0;
-      for (int dy = -1; dy <= 1; ++dy) {
-        for (int dx = -1; dx <= 1; ++dx) {
-          correct += in_buf(x + dx, y + dy);
-        }
-      }
-      ASSERT_EQ(correct, out_buf(x, y)) << x << " " << y;
-    }
-  }
-}
-
-class concatenated_result : public testing::TestWithParam<bool> {};
-
-INSTANTIATE_TEST_SUITE_P(schedule, concatenated_result, testing::Values(false, true));
-
-TEST_P(concatenated_result, pipeline) {
-  bool no_alias_buffers = GetParam();
-  // Make the pipeline
-  node_context ctx;
-
-  auto in1 = buffer_expr::make(ctx, "in1", 2, sizeof(short));
-  auto in2 = buffer_expr::make(ctx, "in2", 2, sizeof(short));
-  auto out = buffer_expr::make(ctx, "out", 2, sizeof(short));
-
-  auto intm1 = buffer_expr::make(ctx, "intm1", 2, sizeof(short));
-  auto intm2 = buffer_expr::make(ctx, "intm2", 2, sizeof(short));
-
-  var x(ctx, "x");
-  var y(ctx, "y");
-
-  // In this pipeline, the result is copied to the output. We should just compute the result directly in the output.
-  func add1 = func::make(add_1<short>, {{{in1, {point(x), point(y)}}}}, {{{intm1, {x, y}}}});
-  func add2 = func::make(add_1<short>, {{{in2, {point(x), point(y)}}}}, {{{intm2, {x, y}}}});
-  func concatenated =
-      func::make_concat({intm1, intm2}, {out, {x, y}}, 1, {0, in1->dim(1).extent(), out->dim(1).extent()});
-
-  pipeline p = build_pipeline(ctx, {in1, in2}, {out}, build_options{.no_alias_buffers = no_alias_buffers});
-
-  // Run the pipeline.
-  const int W = 20;
-  const int H1 = 4;
-  const int H2 = 7;
-  buffer<short, 2> in1_buf({W, H1});
-  buffer<short, 2> in2_buf({W, H2});
-  init_random(in1_buf);
-  init_random(in2_buf);
-
-  buffer<short, 2> out_buf({W, H1 + H2});
-  out_buf.allocate();
-
-  // Not having span(std::initializer_list<T>) is unfortunate.
-  const raw_buffer* inputs[] = {&in1_buf, &in2_buf};
-  const raw_buffer* outputs[] = {&out_buf};
-  test_context eval_ctx;
-  p.evaluate(inputs, outputs, eval_ctx);
-  if (!no_alias_buffers) {
-    ASSERT_EQ(eval_ctx.heap.total_count, 0);
-  }
-
-  for (int y = 0; y < H1 + H2; ++y) {
-    for (int x = 0; x < W; ++x) {
-      ASSERT_EQ(out_buf(x, y), (y < H1 ? in1_buf(x, y) : in2_buf(x, y - H1)) + 1);
-    }
-  }
-
-  // Also visualize this pipeline.
-  check_visualize("concatenate_" + std::to_string(no_alias_buffers) + ".html", p, inputs, outputs, &ctx);
-
-  if (no_alias_buffers == true) {
-    check_replica_pipeline(
-        define_replica_pipeline(ctx, {in1, in2}, {out}, build_options{.no_alias_buffers = no_alias_buffers}));
-  }
-}
-
-class transposed_result : public testing::TestWithParam<std::tuple<bool, std::vector<int>>> {};
-
-INSTANTIATE_TEST_SUITE_P(schedule, transposed_result,
-    testing::Combine(testing::Values(false, true), testing::Values(std::vector<int>{0, 1, 2}, std::vector<int>{0, 2, 1},
-                                                       std::vector<int>{1, 2, 0}, std::vector<int>{0, 0, 0})));
-
-template <typename T>
-std::vector<T> permute(span<const int> p, const std::vector<T>& x) {
-  std::vector<T> result(x.size());
-  for (std::size_t i = 0; i < x.size(); ++i) {
-    result[i] = x[p[i]];
-  }
-  return result;
-}
-
-bool is_permutation(span<const int> p) {
-  std::vector<int> unpermuted(p.size());
-  std::iota(unpermuted.begin(), unpermuted.end(), 0);
-  return std::is_permutation(p.begin(), p.end(), unpermuted.begin());
-}
-
-TEST_P(transposed_result, pipeline) {
-  bool no_alias_buffers = std::get<0>(GetParam());
-  const std::vector<int>& permutation = std::get<1>(GetParam());
-
-  // Make the pipeline
-  node_context ctx;
-
-  auto in = buffer_expr::make(ctx, "in", 3, sizeof(short));
-  auto out = buffer_expr::make(ctx, "out", 3, sizeof(short));
-
-  auto intm = buffer_expr::make(ctx, "intm", 3, sizeof(short));
-
-  var x(ctx, "x");
-  var y(ctx, "y");
-  var z(ctx, "z");
-
-  // In this pipeline, the result is copied to the output. We should just compute the result directly in the output.
-  func add = func::make(add_1<short>, {{{in, {point(x), point(y), point(z)}}}}, {{{intm, {x, y, z}}}});
-  func transposed =
-      func::make_copy({intm, permute<interval_expr>(permutation, {point(x), point(y), point(z)})}, {out, {x, y, z}});
-
-  pipeline p = build_pipeline(ctx, {in}, {out}, build_options{.no_alias_buffers = no_alias_buffers});
-
-  // Run the pipeline.
-  const int W = 20;
-  const int H = 4;
-  const int D = 7;
-  buffer<short, 3> in_buf(permute<index_t>(permutation, {W, H, D}));
-  init_random(in_buf);
-
-  buffer<short, 3> out_buf({W, H, D});
-  out_buf.allocate();
-
-  // Not having span(std::initializer_list<T>) is unfortunate.
-  const raw_buffer* inputs[] = {&in_buf};
-  const raw_buffer* outputs[] = {&out_buf};
-  test_context eval_ctx;
-  p.evaluate(inputs, outputs, eval_ctx);
-  if (is_permutation(permutation) && !no_alias_buffers) {
-    ASSERT_EQ(eval_ctx.heap.total_count, 0);
-  } else {
-    ASSERT_EQ(eval_ctx.heap.total_count, 1);
-  }
-
-  for (int z = 0; z < D; ++z) {
-    for (int y = 0; y < H; ++y) {
-      for (int x = 0; x < W; ++x) {
-        ASSERT_EQ(out_buf(x, y, z), in_buf(permute<index_t>(permutation, {x, y, z})) + 1);
-      }
-    }
-  }
-}
-
-TEST(stacked_result, pipeline) {
-  // Make the pipeline
-  node_context ctx;
-
-  auto in1 = buffer_expr::make(ctx, "in1", 2, sizeof(short));
-  auto in2 = buffer_expr::make(ctx, "in2", 2, sizeof(short));
-  auto out = buffer_expr::make(ctx, "out", 3, sizeof(short));
-
-  auto intm1 = buffer_expr::make(ctx, "intm1", 2, sizeof(short));
-  auto intm2 = buffer_expr::make(ctx, "intm2", 2, sizeof(short));
-
-  var x(ctx, "x");
-  var y(ctx, "y");
-  var z(ctx, "z");
-
-  // In this pipeline, the result is copied to the output. We should just compute the result directly in the output.
-  func add1 = func::make(add_1<short>, {{{in1, {point(x), point(y)}}}}, {{{intm1, {x, y}}}});
-  func add2 = func::make(add_1<short>, {{{in2, {point(x), point(y)}}}}, {{{intm2, {x, y}}}});
-  func stacked = func::make_stack({intm1, intm2}, {out, {x, y, z}}, 2);
-
-  pipeline p = build_pipeline(ctx, {in1, in2}, {out});
-
-  // Run the pipeline.
-  const int W = 20;
-  const int H = 8;
-  buffer<short, 2> in1_buf({W, H});
-  buffer<short, 2> in2_buf({W, H});
-  init_random(in1_buf);
-  init_random(in2_buf);
-
-  buffer<short, 3> out_buf({W, H, 2});
-  out_buf.allocate();
-
-  // Not having span(std::initializer_list<T>) is unfortunate.
-  const raw_buffer* inputs[] = {&in1_buf, &in2_buf};
-  const raw_buffer* outputs[] = {&out_buf};
-  test_context eval_ctx;
-  p.evaluate(inputs, outputs, eval_ctx);
-  ASSERT_EQ(eval_ctx.heap.total_count, 0);
-
-  for (int y = 0; y < H; ++y) {
-    for (int x = 0; x < W; ++x) {
-      ASSERT_EQ(out_buf(x, y, 0), in1_buf(x, y) + 1);
-      ASSERT_EQ(out_buf(x, y, 1), in2_buf(x, y) + 1);
-    }
-  }
-
-  check_replica_pipeline(define_replica_pipeline(ctx, {in1, in2}, {out}));
 }
 
 class padded_stencil : public testing::TestWithParam<int> {};
